requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
loguru>=0.7.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
httpx>=0.25.0
diskcache>=5.6.0
redis>=5.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
jsonschema>=4.20.0
sqlalchemy>=2.0.0
boto3>=1.34.0
anthropic>=0.7.0
openai>=1.3.0
PyYAML>=6.0
python-multipart>=0.0.6
aiofiles>=23.2.0
hypothesis>=6.88.0
mypy>=1.7.0
ruff>=0.1.0
black>=23.11.0
coverage>=7.3.0